Tidy up utils
