+++ lib/python/node_manager/plugins/release.py
@@ -10,6 +10,7 @@ import hou
 from node_manager import utils
 from node_manager.utils import definition as definition_utils
 from node_manager.utils import nodes
+from node_manager.utils import nodetypeutils
 
 plugin_name = "DefaultRelease"
 plugin_class = "release"
@@ -64,7 +65,7 @@ class NodeManagerPlugin(object):
         node_file_path = definition.libraryFilePath()
         if not release_comment:
             release_comment = "Updated {name}".format(
-                name=utils.node_type_name(node_file_path)
+                name=nodetypeutils.node_type_name(node_file_path)
             )
 
         logger.debug("Release comment: {comment}".format(comment=release_comment))
