+++ lib/python/node_manager/plugins/rezrelease.py
@@ -9,14 +9,13 @@ import subprocess
 import time
 from tempfile import mkstemp
 
-from packaging.version import parse
-
 from git import Repo
 from git.exc import NoSuchPathError, InvalidGitRepositoryError
 
 import hou
 
 from node_manager import utils
+from node_manager.utils import nodetypeutils
 from node_manager.plugins import release
 
 plugin_name = "RezRelease"
@@ -395,7 +394,7 @@ class NodeManagerPlugin(release.NodeManagerPlugin):
         node_file_path = definition.libraryFilePath()
         if not release_comment:
             release_comment = "Updated {name}".format(
-                name=utils.node_type_name(node_file_path)
+                name=nodetypeutils.node_type_name(node_file_path)
             )
 
         # Define the release directory
