+++ lib/python/node_manager/utils/__init__.py
@@ -9,6 +9,7 @@ import time
 import hou
 
 from node_manager import manager
+from node_manager.utils import nodetypeutils
 
 
 logger = logging.getLogger(__name__)
@@ -19,6 +20,9 @@ def get_manager():
 
     Returns:
         node_manager.NodeManager: The current instance of Node Manager.
+
+    Raises:
+        RuntimeError: Node Manager not initialised.
     """
     manager_instance = manager.NodeManager.instance
     if not manager_instance:
@@ -103,108 +107,6 @@ def display_message(
     return result
 
 
-def valid_node_type_name(node_type_name):
-    """
-    Validate the nodeTypeName for the given hou.HDADefinition.
-
-    Args:
-        node_type_name(str): The node type name to validate.
-
-    Returns:
-        (bool): Is the node type name valid?
-    """
-    name_components = node_type_name_components(node_type_name)
-    if len(name_components) >= 2:
-        return True
-
-    return False
-
-
-def node_type_name_components(node_type_name):
-    """Get the node type components.
-
-    A hou.HDADefinition nodeTypeName is formated as follows: namespace::name::version.
-    For a given node type name return a list of the various components.
-
-    Args:
-        node_type_name(str): The node type name toy get the components from.
-
-    Returns:
-        (list): A list of name components.
-    """
-    return node_type_name.split("::")
-
-
-def node_type_namespace(node_type_name, new_namespace=None):
-    """Get the node type namespace.
-
-    Get the nodeType namespace from the given nodeTypeName, or use te new namespace if
-    one has been provided.
-
-    Args:
-        node_type_name(str): The node type name to get the namespace from.
-        new_namespace(:obj:`str`,optional): The updated namespace.
-
-    Returns:
-        (str): The namespace extracted from the node type name.
-    """
-    if new_namespace:
-        return new_namespace
-
-    if valid_node_type_name(node_type_name):
-        name_sections = node_type_name_components(node_type_name)
-        namespace_name_sections = name_sections[:-2]
-        return ".".join(namespace_name_sections)
-
-    return None
-
-
-def node_type_name(node_type_name, new_name=None):
-    """Get the node type name.
-
-    Get the nodeType name from the given nodeTypeName, or use te new name if one has
-    been provided.
-
-    Args:
-        node_type_name(str): The node type name to get the name from.
-        new_name(:obj:`str`,optional): The updated name.
-
-    Returns:
-        (str): The name extracted from the node type name.
-    """
-    if new_name:
-        return new_name
-
-    if valid_node_type_name(node_type_name):
-        name_sections = node_type_name_components(node_type_name)
-        return name_sections[-2]
-
-    return node_type_name
-
-
-def node_type_version(node_type_name, new_version=None):
-    """Get the node type version.
-
-    Get the nodeType version from the given nodeTypeName, or use te new version if one
-    has been provided.
-
-    Args:
-        node_type_name(str): The node type name to get the version from.
-        new_version(:obj:`str`,optional): The updated version number.
-
-    Returns:
-        (str): The version extracted from the node type name.
-    """
-    if new_version:
-        return new_version
-
-    if valid_node_type_name(node_type_name):
-        name_sections = node_type_name_components(node_type_name)
-        return name_sections[-1]
-
-    return None
-
-
 def node_type_index(node_type_name, category):
     """Generate a node type index.
 
@@ -219,8 +121,8 @@ def node_type_index(node_type_name, category):
         index(str): The node type index based on the given criteria.
     """
     index = None
-    if valid_node_type_name(node_type_name):
-        name_sections = node_type_name_components(node_type_name)
+    if nodetypeutils.valid_node_type_name(node_type_name):
+        name_sections = nodetypeutils.node_type_name_components(node_type_name)
         name_sections[-2] = "{category}/{name}".format(
             category=category, name=name_sections[-2]
         )
@@ -241,9 +143,9 @@ def release_branch_name(definition):
     """
     category = definition.nodeTypeCategory().name()
     current_name = definition.nodeTypeName()
-    namespace = node_type_namespace(current_name)
-    name = node_type_name(current_name)
-    version = node_type_version(current_name)
+    namespace = nodetypeutils.node_type_namespace(current_name)
+    name = nodetypeutils.node_type_name(current_name)
+    version = nodetypeutils.node_type_version(current_name)
     ts = time.gmtime()
     release_time = time.strftime("%d-%m-%y-%H-%M-%S", ts)
     return "release_{category}-{namespace}{name}-{version}-{time}".format(
@@ -272,9 +174,9 @@ def node_type_name_from_components(definition, namespace=None, name=None, versio
         (str): The updated node type name.
     """
     current_name = definition.nodeTypeName()
-    new_namespace = node_type_namespace(current_name, new_namespace=namespace)
-    new_name = node_type_name(current_name, new_name=name)
-    new_version = node_type_version(current_name, new_version=version)
+    new_namespace = nodetypeutils.node_type_namespace(current_name, new_namespace=namespace)
+    new_name = nodetypeutils.node_type_name(current_name, new_name=name)
+    new_version = nodetypeutils.node_type_version(current_name, new_version=version)
     return "{namespace}::{name}::{version}".format(
         namespace=new_namespace, name=new_name, version=new_version
     )
@@ -299,11 +201,11 @@ def editable_hda_path_from_components(definition, edit_dir, namespace=None, name
     """
     category = definition.nodeTypeCategory()
     current_name = definition.nodeTypeName()
-    if valid_node_type_name(current_name):
+    if nodetypeutils.valid_node_type_name(current_name):
         # If the name is valid, use it
         logger.debug("Using valid node type name: %s", current_name)
-        new_namespace = node_type_namespace(current_name, new_namespace=namespace)
-        new_name = node_type_name(current_name, new_name=name)
+        new_namespace = nodetypeutils.node_type_namespace(current_name, new_namespace=namespace)
+        new_name = nodetypeutils.node_type_name(current_name, new_name=name)
         full_name = "{namespace}{name}".format(
             namespace="{namespace}_".format(namespace=new_namespace) if new_namespace else "",
             name=new_name,
@@ -338,9 +240,9 @@ def expanded_hda_name(definition):
     """
     category = definition.nodeTypeCategory().name()
     current_name = definition.nodeTypeName()
-    name = node_type_name(current_name)
-    namespace = node_type_namespace(current_name)
-    version = node_type_version(current_name)
+    name = nodetypeutils.node_type_name(current_name)
+    namespace = nodetypeutils.node_type_namespace(current_name)
+    version = nodetypeutils.node_type_version(current_name)
     return "{category}_{namespace}{name}.{version}.hda".format(
         category=category,
         namespace="{namespace}.".format(namespace=namespace) if namespace else "",
